feat: Refactor RelayTransport and introduce RelayJsonRpcTransport
- Promoted RelayTransport as the default implementation, renaming from thin-adapter.
- Removed tracing utilities and debug logging to adhere to the "thin" principle.
- Introduced RelayJsonRpcTransport for JSON-RPC interface over HTTP transport.
- Updated exports in index.ts to reflect the new structure.
- Added documentation updates for the changes made.
