+++ .serena/memories/relay_transport_thin_refactoring.md
@@ -1,0 +1,53 @@
+# RelayTransport "薄い実装" リファクタリング
+
+## 実施日: 2025-09-13
+
+## 目的
+
+RelayTransportを「薄い実装」の原則に従って簡素化し、Hatagoの哲学を実現する
+
+## 実施内容
+
+### 削除した機能
+
+1. **トレーシング機能の完全削除**
+   - `packages/transport/src/tracing.ts` ファイル削除（200行）
+   - RelayTransport内のすべてのトレース関連コード削除
+
+2. **デバッグログの削除**
+   - すべての `console.error('[RelayTransport]')` 削除
+   - debugフラグと関連処理を削除
+
+3. **GETリクエストのモック処理削除**
+   - 偽のレスポンスを返す処理を削除
+   - StreamableHTTPTransportに直接委譲
+
+4. **RelayJsonRpcTransportの分離**
+   - 新ファイル `relay-jsonrpc-transport.ts` を作成（79行）
+   - RelayTransportから分離
+
+## 成果
+
+### 行数の削減
+
+- **削減前**: 345行 (relay-transport.ts)
+- **削減後**: 141行 (relay-transport.ts) + 79行 (relay-jsonrpc-transport.ts) = 220行
+- **削減率**: 約60%削減（345行 → 141行）
+
+### 原則への準拠
+
+- ✅ Don't add, remove - 不要な機能を削除
+- ✅ Don't transform, relay - データ変換を最小限に
+- ✅ Don't judge, pass through - 判断ロジックを削除
+- ✅ Don't thicken, stay thin - 薄い実装を維持
+
+### コードの改善
+
+- StreamableHTTPTransportの薄いラッパーとして機能
+- メソッドは単純に転送するだけ
+- 透明性の向上
+- メンテナンス性の向上
+
+## 注記
+
+理想的な50行には届かなかったが、インターフェースの互換性を保ちながら141行まで削減できた。これは現実的な「薄い実装」として十分な成果。
